mapper, exception handler, validation

+++ src/main/java/am/itspace/authorbookrest/service/impl/AuthorServiceImpl.java
@@ -1,9 +1,10 @@
 package am.itspace.authorbookrest.service.impl;
 
-import am.itspace.authorbookrest.converter.AuthorConverter;
 import am.itspace.authorbookrest.dto.AuthorDto;
 import am.itspace.authorbookrest.dto.SaveAuthorRequest;
 import am.itspace.authorbookrest.entity.Author;
+import am.itspace.authorbookrest.exception.AuthorNotFoundException;
+import am.itspace.authorbookrest.mapper.AuthorMapper;
 import am.itspace.authorbookrest.repository.AuthorRepository;
 import am.itspace.authorbookrest.service.AuthorService;
 import lombok.RequiredArgsConstructor;
@@ -12,7 +13,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -22,16 +22,12 @@ import java.util.Optional;
 public class AuthorServiceImpl implements AuthorService {
 
     private final AuthorRepository authorRepository;
-    private final AuthorConverter authorConverter;
+    private final AuthorMapper authorMapper;
 
     @Override
     public List<AuthorDto> findAll() {
         List<Author> authors = authorRepository.findAll();
-        List<AuthorDto> result = new ArrayList<>();
-        for (Author author : authors) {
-            result.add(authorConverter.fromEntityToDto(author));
-        }
-        return result;
+        return authorMapper.toDtoList(authors);
     }
 
     @Override
@@ -41,13 +37,16 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public AuthorDto save(SaveAuthorRequest authorRequest) {
-        Author author = authorRepository.save(authorConverter.fromDtoToEntity(authorRequest));
+        Author author = authorRepository.save(authorMapper.toEntity(authorRequest));
 
-        return authorConverter.fromEntityToDto(author);
+        return authorMapper.toDto(author);
     }
 
     @Override
     public void deleteById(int id) {
+        if (!authorRepository.existsById(id)) {
+            throw new AuthorNotFoundException("Author not found with " + id + " id");
+        }
         authorRepository.deleteById(id);
     }
 
@@ -56,10 +55,10 @@ public class AuthorServiceImpl implements AuthorService {
         Author author = authorRepository
                 .findById(id)
                 .orElse(null);
-        if(author == null){
+        if (author == null) {
             return null;
         }
-        return authorConverter.fromEntityToDto(author);
+        return authorMapper.toDto(author);
     }
 
     @Override
